==> 0-id-prep.ipynb <==
# ID Prep

Preparation for student ids. 
- Download IDs from moodle (admin block > gear > responses > download as .csv)


## Raw ID from moodle form

raw_id = pd.read_csv('./id-raw.csv')

raw_id.head()

# remove and rename columns
id_clean = raw_id.drop(['State','Started on','Completed','Time taken', 'Grade/3.00'],axis=1).rename(columns={'Response 1':'Matrikel','Response 2':'SS','Response 3':'PO'})

id_clean.head()

# export the id-table with removed and renamed columns to csv
id_clean.to_csv('./id_clean.csv',index=False)

## Update the ID table manually with data from email

# update id_clean manually
id_clean_updated = pd.read_csv('./id-clean-updated.csv')

# merge two columns names into one
id_clean_updated['Name'] = id_clean_updated['First name'] + ' ' + id_clean_updated['Surname']

# drop two redundant name columns and rename matrikel column
id_clean_updated = id_clean_updated.drop(['Surname','First name'],axis=1).rename(columns={'Matrikelnummer':'Matrikel'})

id_clean_updated.head()

id_clean_updated.info()

## Remove duplicate rows (by name or matrikel)

dups_by_name = id_clean_updated[id_clean_updated.duplicated(['Name'])]

dups_by_name.head()

id_clean_updated_nodups = id_clean_updated.drop(dups_by_name.index,axis=0)

id_clean_updated_nodups.info()

# contains dups by matrikel and '-' values
dups_by_matrikel = id_clean_updated_nodups[id_clean_updated_nodups.duplicated(['Matrikel'])]

dups_by_matrikel.head()

id_clean_updated_nodups2 = id_clean_updated_nodups.drop(dups_by_matrikel.index,axis=0)

id_clean_updated_nodups2.info()

# drop last row wih '-' value
id_clean_updated_nodups3 = id_clean_updated_nodups2.drop(id_clean_updated_nodups2[id_clean_updated_nodups2['Matrikel'] == '-'].index,axis=0)

id_clean_updated_nodups3.head()

id_clean_updated_nodups3.info()

id_clean_updated.to_csv('./id_clean_updated_merged.csv',index=False)

# use this table in the validator script
id_clean_updated_nodups3.to_csv('./id_clean_updated_nodups.csv',index=False)
